apps/api_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
python-dotenv>=1.0.1

llama-index>=0.12.0
llama-index-vector-stores-qdrant>=0.8.0
qdrant-client>=1.13.0

llama-index-llms-ollama>=0.5.0
llama-index-embeddings-huggingface>=0.5.0
sentence-transformers>=3.4.0
